Remove language referring to "subprojects"
The notion of a "subproject" is not clearly defined in the governance
rules, and based on the existing minimal usage in the ROLES document
seems to correspond to a Working Group. This change removes language
referring to subprojects in favor of consolidating on only Working
Groups as the governance structure below the TOC.

Sadly, this change will conflict with the excellent PR #60.