Archive silk repo, it was merged into silk-release
We had originally setup silk in an enitre other repo with the intention
that it could be imported, used, and extended by other projects.
This never happened and now we have a separate repo for `silk` when it
is only used in `silk-release.

We have moved `silk` into `silk-release` and therefore `silk` is no
longer necessary.

More details:
- https://github.com/cloudfoundry/silk/issues/52
- https://github.com/cloudfoundry/silk-release/pull/89

[#185149418](https://www.pivotaltracker.com/story/show/185149418)