Archive repository bosh-package-java-release
Based on the discussions in https://github.com/cloudfoundry/community/pull/653 the decision was taken to
archive the `bosh-package-java-release` repository. The only user of it is the App Autoscaler Area from
the ARI WG but they are in favour to [package Java themself](https://github.com/cloudfoundry/community/pull/653#issuecomment-1723409747) instead of maintaining `bosh-package-java-release` repository.