Add blgm as approver to app-runtime-interfaces
I received an email saying that due to some cleanup I was going to have my access to [jsonry](https://github.com/cloudfoundry/jsonry) removed. As the person who maintains this repo, I thought that could be unhelpful as it could become unmaintained.

JSONry was a spin-off from some code that once lived in the CF CLI, and it seems that it's [actually used as a library](https://github.com/cloudfoundry/jsonry/network/dependents) by other projects.

In addition to maintaining JSONry, I have made significant contributions to the [CF CLI](https://github.com/cloudfoundry/cli/commits?author=blgm), so being an Approver in this group could make sense.

I'm completely behind the idea to clean up authorization and ownership of repos. If it's felt that I don't quite meet the criteria, or that they would be better solutions, then I don't have a problem with that. My aim is to be helpful by continuing to help out with JSONry and the services sub-commands in the CF CLI, but there are many ways to achieve that.