Correct start and end dates of the 2024 CFF TOC and Paketo SC elections
This fixes the UTC start and end timestamps for the elections to match the PDT
timestamps in the comments. This start date and time also then allows for
nominations through the entirety of the Pacific Time day on Jun 7th, as seems
to be the intent based on the [announcement message on the cf-dev mailing list](https://lists.cloudfoundry.org/g/cf-dev/message/9479).