Remove user ifindlay-cci
The [Sync Github Organization
Settings](https://github.com/cloudfoundry/community/actions/runs/16989865344/job/48166497010)
is failing because there is a issue with the github user ifindlay-cci.
This pr suggest to remove the user until we clarify the issue and the
change can be reverted when we have clarity on it and it doesn't fail
the automation.