Reduce the request batch size in inactive user request
According to [this
discussion](https://github.com/orgs/community/discussions/24844#discussioncomment-3245605)
there are 10s timeouts on GitHub side for all API calls and the
requested amound of data could be a reasond for request to fails because
of that timeout. That is why this change is redecuing the requested size
of 50 users to 20. Testing locally worked and the calls didn't fails
with 502.