Remove Inactive Users FI WG
According to the rules for inactivity defined in [RFC-0025](https://github.com/cloudfoundry/community/blob/main/toc/rfc/rfc-0025-define-criteria-and-removal-process-for-inactive-members.md), the following users should be removed from ARI approvers/reviewers:
@bgandon
@max-soe
@ichandrabhatta
@yharish991
@nouseforaname
@pivotalgeorge
@tas-operability-bot
@cunnie

According to the [revocation policy in the RFC](https://github.com/cloudfoundry/community/blob/main/toc/rfc/rfc-0025-define-criteria-and-removal-process-for-inactive-members.md#remove-the-membership-to-the-cloud-foundry-github-organization), users have two weeks to refute this revocation, if they wish by commenting on this pull-request and open a new pull-request to be re-added as approver after this one is merged. Please also state a reason why you would like to be re-added.
Inactivity was detected by a github action, see #1318

I took @joefitzgerald and @colins out based on the discussions in https://github.com/cloudfoundry/community/pull/1271

@yharish991 could you please state a reason why you and @tas-operability-bot should presered the role?