Remove Inactive Users FI WG
According to the rules for inactivity defined in RFC-0025, the following users should be removed from ARI approvers/reviewers:
@cf-bosh-ci-bot
@torsten-sap

According to the revocation policy in the RFC, users have two weeks to refute this revocation, if they wish by commenting on this pull-request and open a new pull-request to be re-added as approver after this one is merged. Please also state a reason why you would like to be re-added. Inactivity was detected by a github action, see #1318